+++ pdf/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.3.4
 * Add available dimensions for PdfPageFormat
 * Add Document properties
+* Add Page.orientation to force landscape or portrait
 
 # 1.3.3
 * Fix a bug with the RichText Widget